% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CRISPRhh}
\alias{CRISPRhh}
\title{Run a single simulation of (S1) in Domingues, Kuijper et al.
Notation may differ from what is in the main manuscript 
P is the non-immune host that accepts all foreign genetic elements
C is the CRISPR-immune host that accepts only G2 foreign genetic elements
most importantly, G1 is a worse foreign genetic element than G2}
\usage{
CRISPRhh(
  kappa = 0.001,
  gammaPG1 = 1,
  gammaPG2 = 1,
  gammaCG1 = 1,
  gammaCG2 = 1,
  dSP = 5,
  dSC = 5,
  dPG1 = 5,
  dPG2 = 1,
  dCG1 = 5,
  dCG2 = 1,
  psiG1 = 10,
  psiG2 = 10,
  FG1 = 6,
  FG2 = 10,
  FG1G2 = 0,
  pi = 0.5,
  c = 0.02,
  sigma = 0,
  init_S = 50,
  init_fraction_SC = 0.5,
  init_PG1 = 1,
  init_PG2 = 1,
  init_CG1 = 1,
  init_CG2 = 1,
  demog_feedback = TRUE,
  debug = FALSE,
  eul = 0.01,
  max_time = 500000000L
)
}
\arguments{
\item{kappa}{strength of density-dependence}

\item{gammaPG1}{FGE loss rate in P (promiscuous) hosts of G1 FGEs}

\item{gammaPG2}{FGE loss rate in P (promiscuous) hosts of G2 FGEs}

\item{gammaCG1}{FGE loss rate in C (choosy) hosts of G1 FGEs}

\item{gammaCG2}{FGE loss rate in C (choosy) hosts of G2 FGEs}

\item{dSP}{mortality rate of susceptible P (promiscuous) hosts}

\item{dSC}{mortality rate of susceptible C (choosy) hosts}

\item{dPG1}{mortality rate in P (promiscuous) hosts of G1 FGEs}

\item{dPG2}{mortality rate in P (promiscuous) hosts of G2 FGEs}

\item{dCG1}{mortality rate in C (choosy) hosts of G1 FGEs}

\item{dCG2}{mortality rate in C (choosy) hosts of G2 FGEs}

\item{psiG1}{force of infection by G1 FGEs}

\item{psiG2}{force of infection by G2 FGEs}

\item{FG1}{fecundity benefit of the G1 FGE}

\item{FG2}{fecundity benefit of the G2 FGE}

\item{FG1G2}{fecundity benefit of the G1 + G2 FGEs when superinfected}

\item{pi}{probability of resisting infection by the G1 FGE}

\item{c}{cost of resistance}

\item{sigma}{probability of superinfection}

\item{init_S}{initial density of S}

\item{init_fraction_SC}{initial fraction of S that has the C allele, remainder has the P allele}

\item{init_PG1}{initial density of P hosts infected with G1}

\item{init_PG2}{initial density of P hosts infected with G2}

\item{init_CG1}{initial density of C hosts infected with G1}

\item{init_CG2}{initial density of C hosts infected with G2}

\item{demog_feedback}{boolean allowing for demographical feedbacks or not}

\item{debug}{boolean whether to give debug output or not}

\item{eul}{Euler's constant}

\item{max_time}{The maximum number of time steps the numerical simulation shoujld run for}
}
\value{
A \code{data.frame} that contains the various frequencies
}
\description{
Run a single simulation of (S1) in Domingues, Kuijper et al.
Notation may differ from what is in the main manuscript 
P is the non-immune host that accepts all foreign genetic elements
C is the CRISPR-immune host that accepts only G2 foreign genetic elements
most importantly, G1 is a worse foreign genetic element than G2
}
