\name{CRISPRhitchhike-package}
\alias{CRISPRhitchhike-package}
\alias{CRISPRhitchhike}
\docType{package}
\title{
  Epidemiological model of how CRISPR immunity can discriminate between
  good vs better phage / plasmid infections. See Domingues et al 
}
\description{
    The function CRISPRhh parameterizes the model that is used and allows
    one to assess 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Bram Kuijper

Maintainer: Bram Kuijper, a.l.w.kuijper@exeter.ac.uk
}
